#27 move ViewToStringRendererService to Sitko.Core.App.Web
